agregando jornada y terminando metodo

+++ salidas/views.py
@@ -293,44 +293,21 @@ def my_information(request):
     form = MyInformation(request.POST or None)
     signature = TeachersSignature2(request.FILES or None)
 
-    print(teacher.name)
-
-    print(form.is_valid())
-    print(request.FILES)
-    if request.method == 'POST' and len(request.FILES) != 0:
+    if request.method == 'POST' and form.is_valid() and len(request.FILES) != 0:
         email = form.cleaned_data['email']
         jornada = form.cleaned_data['jornada']
         sign = request.FILES['sign']
 
         sign_extention =  sign.__str__().split(".")[1]
-        _5K = 5 * 1024 * 1024 - 1024 # -1K for error
-        if form.is_valid() and (sign_extention == "jpg" or sign_extention == "jpeg" or sign_extention == "png") and sign._size < _5K:
+        size_of_sign = 1 * 1024 * 1024 + 1024 # 1K for error
+        if form.is_valid() and (sign_extention == "jpg" or sign_extention == "jpeg" or sign_extention == "png") and sign._size <= size_of_sign:
+            teacher.mail = email
+            teacher.working_day = jornada
             teacher.signature.delete()
             teacher.signature = sign
             teacher.save()
-
-        teacher = Teacher.objects.get(pk = request.user.id)
-        teacher.mail = email
-        # teacher.working_day = jornada
-        teacher.signature = request.FILES['sign']
-        teacher.save()
-
-    '''
-    teacher_signature = TeacherSignatureForm(request.FILES or None)
-
-    # signature verification and update
-    try:
-        asignature = request.FILES['signature']
-        extention = asignature.__str__().split(".")[1]
-        if extention == "jpg" or extention == "jpeg" or extention == "png":
-            prof.signature.delete()
-            prof.signature = asignature
-            prof.save()
         else:
-            messages.error(request, "Formato de firma invalido. Usar '.jpg' o '.png'.")
-    except:
-        asignature=None
-    '''
+            print("error en validacion de archivos")
 
     return render_to_response("Professor/my_information.html", locals(), context_instance=RequestContext(request))
